requirements.txt
accelerate==0.27.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
docker-pycreds==0.4.0
editdistance==0.6.0
einops==0.8.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
fire==0.6.0
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.19.4
idna==3.6
importlib_metadata==7.0.2
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.22
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mkl-service==2.4.1
# mkl_fft==1.3.8
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.1
numpy==1.24.3
nvidia-ml-py==12.535.133
nvitop==1.3.2
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
patsy==0.5.6
pexpect==4.9.0
pillow==9.4.0
platformdirs==4.2.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pudb==2024.1
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.12
pyparsing==3.1.1
pysocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.4
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.11.3
sconf==0.2.5
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.96
sentry-sdk==1.40.6
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.12
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.2
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.3.0
tiktoken==0.6.0
timm==0.5.4
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.40.1
transformers-stream-generator==0.0.5
typer==0.10.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
urwid==2.6.8
urwid_readline==0.14
uvicorn==0.29.0
wandb==0.12.15
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zss==1.2.0
